project1/requirements.txt
flask==2.3.3
pyjwt==2.8.0
werkzeug==2.3.7 